requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
chromadb==0.5.3
sentence-transformers==3.0.1
httpx==0.27.2
python-dotenv==1.0.1
